changes the float data type to decimal

+++ src/app/models/Product.js
@@ -7,7 +7,7 @@ class Product extends Model {
         name: Sequelize.STRING,
         description: Sequelize.STRING,
         category: Sequelize.STRING,
-        price: Sequelize.FLOAT
+        price: Sequelize.DECIMAL
       },
       {
         sequelize
